2024/09/21  refactor to be consistent

+++ transaction-service/src/main/java/com/jack/transactionservice/client/WalletServiceClient.java
@@ -5,11 +5,13 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 
+import java.math.BigDecimal;
+
 @FeignClient(name = "wallet-service", url = "${WALLET_SERVICE_URL:https://wallet-service:8082}")
 public interface WalletServiceClient {
 
     @PostMapping("/api/wallet/{userId}")
     void updateWalletBalance(@PathVariable("userId") Long userId,
-                             @RequestParam("usdBalance") double usdBalance,
-                             @RequestParam("btcBalance") double btcBalance);
+                             @RequestParam("usdBalance") BigDecimal usdBalance,
+                             @RequestParam("btcBalance") BigDecimal btcBalance);
 }
